Upgraded git sync to keep modified timestamps better

+++ Mutagen.Bethesda.GitSync/GitConversionUtility.cs
@@ -1,5 +1,6 @@
 ﻿using Noggog;
 using Noggog.Utility;
+using Splat;
 using System;
 using System.Collections.Generic;
 using System.IO;
@@ -9,8 +10,9 @@ using System.Threading.Tasks;
 
 namespace Mutagen.Bethesda.GitSync
 {
-    public class GitConversionUtility
+    public class GitConversionUtility : IEnableLogger
     {
+        private static GitConversionUtility Instance = new GitConversionUtility();
         public enum Error { None, DidNotExist, ModKey, Corrupted }
 
         public static async Task<Error> ConvertToBinary<M>(
@@ -48,21 +50,30 @@ namespace Mutagen.Bethesda.GitSync
                         return Error.Corrupted;
                     }
                 }
-                if (!string.IsNullOrWhiteSpace(backupFolder?.Path)
-                    && backupFolder.Value.Exists)
+                if (binaryTargetPath.Exists)
                 {
-                    var backupTarget = new DirectoryPath(Path.Combine(backupFolder.Value.Path, DateTime.Now.ToString("MM-dd-yyyy HH-mm-ss")));
-                    backupTarget.Create();
-                    File.Move(binaryTargetPath.Path, Path.Combine(backupTarget.Path, binaryTargetPath.Name));
-                    CleanBackups(backupFolder.Value.Path, numBackups);
-                }
-                else
-                {
-                    binaryTargetPath.Delete();
+                    // Short circuit if everything already equal
+                    if (FileComparison.FilesAreEqual(targetTempPath.Path, binaryTargetPath.Path))
+                    {
+                        return Error.None;
+                    }
+                    // Make backup
+                    if (!string.IsNullOrWhiteSpace(backupFolder?.Path)
+                        && backupFolder.Value.Exists)
+                    {
+                        var backupTarget = new DirectoryPath(Path.Combine(backupFolder.Value.Path, DateTime.Now.ToString("MM-dd-yyyy HH-mm-ss")));
+                        backupTarget.Create();
+                        File.Move(binaryTargetPath.Path, Path.Combine(backupTarget.Path, binaryTargetPath.Name));
+                        CleanBackups(backupFolder.Value.Path, numBackups);
+                    }
+                    else
+                    {
+                        binaryTargetPath.Delete();
+                    }
                 }
                 File.Move(targetTempPath.Path, binaryTargetPath.Path);
             }
-            
+
             return Error.None;
         }
 
@@ -87,8 +98,48 @@ namespace Mutagen.Bethesda.GitSync
             using (var tmpFolder = new TempFolder(deleteAfter: true))
             {
                 var targetTempPath = new DirectoryPath(Path.Combine(tmpFolder.Dir.Path, "Export"));
+                targetTempPath.Create();
+                bool doingBackup = !string.IsNullOrWhiteSpace(backupFolder?.Path)
+                    && xmlFolderTargetPath.Exists;
+
+                void CopyFolder(string source_dir, string destination_dir)
+                {
+                    foreach (string dir in System.IO.Directory.GetDirectories(source_dir, "*", System.IO.SearchOption.AllDirectories))
+                    {
+                        try
+                        {
+                            System.IO.Directory.CreateDirectory(System.IO.Path.Combine(destination_dir, dir.Substring(source_dir.Length + 1)));
+                        }
+                        catch (Exception ex)
+                        {
+                            Instance.Log().Warn("Exception creating directory during temp copy over: " + ex);
+                        }
+                    }
+
+                    foreach (string file_name in System.IO.Directory.GetFiles(source_dir, "*", System.IO.SearchOption.AllDirectories))
+                    {
+                        var targetPath = System.IO.Path.Combine(destination_dir, file_name.Substring(source_dir.Length + 1));
+                        try
+                        {
+                            System.IO.File.Copy(file_name, targetPath);
+                        }
+                        catch (Exception ex)
+                        {
+                            Instance.Log().Warn($"Exception copying {file_name} to {targetPath} during temp copy over: " + ex);
+                        }
+                    }
+                }
+
+                // Move existing XML to preserve date modified information, etc
+                if (xmlFolderTargetPath.Exists)
+                {
+                    CopyFolder(xmlFolderTargetPath.Path, targetTempPath.Path);
+                }
+
+                // Write XML folder in temp location
                 var binInMod = await instr.CreateBinary(binaryPath);
                 await instr.WriteXmlFolder(binInMod, targetTempPath);
+
                 if (checkCorrectness)
                 {
                     var rexportPath = Path.Combine(tmpFolder.Dir.Path, "Reexport");
@@ -101,17 +152,28 @@ namespace Mutagen.Bethesda.GitSync
                         return Error.Corrupted;
                     }
                 }
-                if (!string.IsNullOrWhiteSpace(backupFolder?.Path)
-                    && xmlFolderTargetPath.Exists)
+
+                if (xmlFolderTargetPath.Exists)
                 {
-                    var backupTarget = new DirectoryPath(Path.Combine(backupFolder.Value.Path, DateTime.Now.ToString("MM-dd-yyyy HH-mm-ss")));
-                    backupTarget.Directory.Create();
-                    Directory.Move(xmlFolderTargetPath.Path, backupTarget.Path);
-                    CleanBackups(backupFolder.Value.Path, numBackups);
-                }
-                else
-                {
-                    xmlFolderTargetPath.DeleteEntireFolder();
+                    // Short circuit if everything already equal
+                    if (xmlFolderTargetPath.Exists
+                        && FileComparison.FoldersAreEqual(targetTempPath.Path, xmlFolderTargetPath.Path))
+                    {
+                        return Error.None;
+                    }
+
+                    // Make backup
+                    if (doingBackup)
+                    {
+                        var backupTarget = new DirectoryPath(Path.Combine(backupFolder.Value.Path, DateTime.Now.ToString("MM-dd-yyyy HH-mm-ss")));
+                        backupTarget.Directory.Create();
+                        Directory.Move(xmlFolderTargetPath.Path, backupTarget.Path);
+                        CleanBackups(backupFolder.Value.Path, numBackups);
+                    }
+                    else
+                    {
+                        xmlFolderTargetPath.DeleteEntireFolder();
+                    }
                 }
                 Directory.Move(targetTempPath.Path, xmlFolderTargetPath.Path);
             }
